Revert "refactor(mcp): consolidate MCP client code into mcp/ module"
This reverts commit 60fbf66bdb9bee6092819f986f065122745947d2.
