+++ crates/aura-web-server/src/streaming/handlers.rs
@@ -1552,7 +1552,7 @@ mod tests {
     }
 
     /// An MCP server's error reaches `handle_tool_result` with the prefix
-    /// `aura::mcp::response` applies, and must be reported as a failure rather
+    /// `aura::mcp_response` applies, and must be reported as a failure rather
     /// than a successful result carrying the error text.
     #[test]
     fn test_handle_tool_result_reports_mcp_error_as_failure() {
