+++ crates/aura/src/mcp.rs
@@ -1,10 +1,305 @@
-use crate::config::McpServerConfig;
-use crate::error::BuilderError;
-use crate::mcp::client::McpClient;
-use rig::completion::ToolDefinition;
+use crate::{config::McpServerConfig, error::BuilderError, mcp_streamable_http::McpClient};
+use futures::{StreamExt, stream::BoxStream};
+use rig::{completion::ToolDefinition, tool::Tool as RigTool};
+use rmcp::transport::streamable_http_client::StreamableHttpClient;
 use serde_json::Value;
+use sse_stream::{Sse, SseStream};
 use std::collections::HashMap;
+use std::sync::{Arc, Mutex};
 use tracing::{debug, info, warn};
+
+/// Simple error type for tool execution
+#[derive(Debug)]
+pub struct ToolExecutionError {
+    message: String,
+}
+
+impl std::fmt::Display for ToolExecutionError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Tool execution error: {}", self.message)
+    }
+}
+
+impl std::error::Error for ToolExecutionError {}
+
+impl From<String> for ToolExecutionError {
+    fn from(message: String) -> Self {
+        Self { message }
+    }
+}
+
+impl From<&str> for ToolExecutionError {
+    fn from(message: &str) -> Self {
+        Self {
+            message: message.to_string(),
+        }
+    }
+}
+
+/// Custom HTTP client that captures the underlying HTTP status when a request
+/// fails.
+///
+/// rmcp's streamable-HTTP transport runs in a background worker: when a request
+/// fails (e.g. 404/401), the worker logs the `reqwest` error and closes the
+/// channel, so `serve_client` only sees "channel closed" — the status code is
+/// lost. By implementing `StreamableHttpClient` ourselves we record the status
+/// into `first_error` *at the layer it occurs*, then read it back after the
+/// connection fails to produce a precise reason.
+#[derive(Clone, Default)]
+pub struct CustomHttpClient {
+    client: reqwest::Client,
+    /// First failing HTTP status observed on this client, if any. Shared across
+    /// clones (the transport worker clones the client) via `Arc`.
+    first_error: Arc<Mutex<Option<String>>>,
+}
+
+impl CustomHttpClient {
+    /// Wrap an existing `reqwest::Client` (already carrying any forwarded
+    /// headers, including auth) so transport HTTP errors can be captured.
+    pub fn from_reqwest(client: reqwest::Client) -> Self {
+        Self {
+            client,
+            first_error: Arc::new(Mutex::new(None)),
+        }
+    }
+
+    /// Shared handle to the first captured failing HTTP status, if any.
+    pub fn first_error(&self) -> Arc<Mutex<Option<String>>> {
+        Arc::clone(&self.first_error)
+    }
+
+    /// Record the first non-success HTTP status seen (first error wins, so the
+    /// root cause isn't overwritten by any follow-on failures).
+    fn record_http_status(&self, status: reqwest::StatusCode) {
+        if status.is_success() {
+            return;
+        }
+        if let Ok(mut guard) = self.first_error.lock()
+            && guard.is_none()
+        {
+            *guard = Some(format!("{}{status}", aura_events::HTTP_STATUS_MARKER));
+        }
+    }
+}
+
+impl StreamableHttpClient for CustomHttpClient {
+    type Error = reqwest::Error;
+
+    async fn get_stream(
+        &self,
+        uri: Arc<str>,
+        session_id: Arc<str>,
+        last_event_id: Option<String>,
+        _auth_token: Option<String>, // auth flows through the client's default headers
+    ) -> Result<
+        BoxStream<'static, Result<Sse, sse_stream::Error>>,
+        rmcp::transport::streamable_http_client::StreamableHttpError<Self::Error>,
+    > {
+        use reqwest::header::ACCEPT;
+        use rmcp::transport::common::http_header::{
+            EVENT_STREAM_MIME_TYPE, HEADER_LAST_EVENT_ID, HEADER_SESSION_ID, JSON_MIME_TYPE,
+        };
+
+        let mut request_builder = self
+            .client
+            .get(uri.as_ref())
+            .header(ACCEPT, "application/json, text/event-stream")
+            .header(HEADER_SESSION_ID, session_id.as_ref());
+
+        if let Some(last_event_id) = last_event_id {
+            request_builder = request_builder.header(HEADER_LAST_EVENT_ID, last_event_id);
+        }
+
+        let response = request_builder
+            .send()
+            .await
+            .map_err(rmcp::transport::streamable_http_client::StreamableHttpError::Client)?;
+        if response.status() == reqwest::StatusCode::METHOD_NOT_ALLOWED {
+            // Not a failure — the server just doesn't support the SSE GET.
+            return Err(rmcp::transport::streamable_http_client::StreamableHttpError::ServerDoesNotSupportSse);
+        }
+        self.record_http_status(response.status());
+        let response = response
+            .error_for_status()
+            .map_err(rmcp::transport::streamable_http_client::StreamableHttpError::Client)?;
+
+        match response.headers().get(reqwest::header::CONTENT_TYPE) {
+            Some(ct) => {
+                // Accept both `text/event-stream` and `application/json`, matching
+                // rmcp's reference reqwest client — a server may answer the GET
+                // stream with either. Rejecting JSON here would mark an otherwise
+                // healthy server as `Failed`.
+                if !ct.as_bytes().starts_with(EVENT_STREAM_MIME_TYPE.as_bytes())
+                    && !ct.as_bytes().starts_with(JSON_MIME_TYPE.as_bytes())
+                {
+                    return Err(rmcp::transport::streamable_http_client::StreamableHttpError::UnexpectedContentType(Some(
+                        String::from_utf8_lossy(ct.as_bytes()).to_string(),
+                    )));
+                }
+            }
+            None => {
+                return Err(rmcp::transport::streamable_http_client::StreamableHttpError::UnexpectedContentType(None));
+            }
+        }
+
+        let event_stream = SseStream::from_byte_stream(response.bytes_stream()).boxed();
+        Ok(event_stream)
+    }
+
+    async fn delete_session(
+        &self,
+        uri: Arc<str>,
+        session: Arc<str>,
+        _auth_token: Option<String>, // auth flows through the client's default headers
+    ) -> Result<(), rmcp::transport::streamable_http_client::StreamableHttpError<Self::Error>> {
+        use rmcp::transport::common::http_header::HEADER_SESSION_ID;
+
+        let response = self
+            .client
+            .delete(uri.as_ref())
+            .header(HEADER_SESSION_ID, session.as_ref())
+            .send()
+            .await
+            .map_err(rmcp::transport::streamable_http_client::StreamableHttpError::Client)?;
+
+        if response.status() == reqwest::StatusCode::METHOD_NOT_ALLOWED {
+            return Err(rmcp::transport::streamable_http_client::StreamableHttpError::ServerDoesNotSupportDeleteSession);
+        }
+        response
+            .error_for_status()
+            .map_err(rmcp::transport::streamable_http_client::StreamableHttpError::Client)?;
+        Ok(())
+    }
+
+    async fn post_message(
+        &self,
+        uri: Arc<str>,
+        message: rmcp::model::ClientJsonRpcMessage,
+        session_id: Option<Arc<str>>,
+        _auth_token: Option<String>, // auth flows through the client's default headers
+    ) -> Result<
+        rmcp::transport::streamable_http_client::StreamableHttpPostResponse,
+        rmcp::transport::streamable_http_client::StreamableHttpError<Self::Error>,
+    > {
+        use rmcp::transport::common::http_header::{HEADER_SESSION_ID, JSON_MIME_TYPE};
+
+        // Approver header overrides ride the request's extensions: present
+        // only on the one gated call whose approval captured them, and
+        // never serialized into the JSON body below (the rmcp serializer
+        // emits only `_meta`).
+        let approver_overrides = crate::approver_headers::extract_from_client_message(&message);
+
+        let mut request_builder = self
+            .client
+            .post(uri.as_ref())
+            .header(reqwest::header::CONTENT_TYPE, JSON_MIME_TYPE)
+            .header(
+                reqwest::header::ACCEPT,
+                "application/json, text/event-stream",
+            );
+
+        // Forward the negotiated session id on every post after `initialize`.
+        // The server returns the id in the initialize response and requires it
+        // on subsequent requests (`notifications/initialized`, tool calls, …);
+        // dropping it makes the server reject them (FastMCP: 400, rmcp: 422),
+        // which the transport worker then collapses into "channel closed".
+        if let Some(session_id) = session_id {
+            request_builder = request_builder.header(HEADER_SESSION_ID, session_id.as_ref());
+        }
+
+        // Per-request override headers beat the client's frozen
+        // `default_headers` for exactly this request. Applied last, after
+        // every transport-owned header above; reserved names are
+        // additionally rejected at config parse, so framing and session
+        // routing stay intact.
+        let mut request_builder = request_builder.json(&message);
+        if let Some(overrides) = approver_overrides {
+            request_builder = overrides.apply_to(request_builder);
+        }
+
+        let response = request_builder
+            .send()
+            .await
+            .map_err(rmcp::transport::streamable_http_client::StreamableHttpError::Client)?;
+        // Capture the status before error_for_status consumes the response — this
+        // is the initialize/JSON-RPC POST, where auth (401) and endpoint (404)
+        // failures surface, and where the transport worker would otherwise hide
+        // them behind a "channel closed" error.
+        let status = response.status();
+        self.record_http_status(status);
+        let response = response
+            .error_for_status()
+            .map_err(rmcp::transport::streamable_http_client::StreamableHttpError::Client)?;
+
+        // A notification/response-less POST (e.g. `notifications/initialized`)
+        // comes back as 202 Accepted / 204 No Content with an empty body. Some
+        // servers (FastMCP) still tag the empty body `application/json`; parsing
+        // it as a JSON-RPC message fails and the worker reports "channel closed".
+        // Short-circuit on these statuses before touching the body, matching
+        // rmcp's reference reqwest client.
+        if matches!(
+            status,
+            reqwest::StatusCode::ACCEPTED | reqwest::StatusCode::NO_CONTENT
+        ) {
+            return Ok(
+                rmcp::transport::streamable_http_client::StreamableHttpPostResponse::Accepted,
+            );
+        }
+
+        // Extract session ID from headers before consuming response
+        let session_id = response
+            .headers()
+            .get("mcp-session-id")
+            .and_then(|h| h.to_str().ok())
+            .map(|s| s.to_string());
+
+        // Snapshot the content type as an owned string before consuming the
+        // response body in the branches below.
+        let content_type = response
+            .headers()
+            .get(reqwest::header::CONTENT_TYPE)
+            .map(|ct| String::from_utf8_lossy(ct.as_bytes()).to_string());
+
+        match content_type.as_deref() {
+            Some(ct) if ct.starts_with("application/json") => {
+                let json_text = response.text().await.map_err(
+                    rmcp::transport::streamable_http_client::StreamableHttpError::Client,
+                )?;
+                let json_message: rmcp::model::ServerJsonRpcMessage =
+                    serde_json::from_str(&json_text).map_err(
+                        rmcp::transport::streamable_http_client::StreamableHttpError::Deserialize,
+                    )?;
+                Ok(
+                    rmcp::transport::streamable_http_client::StreamableHttpPostResponse::Json(
+                        json_message,
+                        session_id,
+                    ),
+                )
+            }
+            Some(ct) if ct.starts_with("text/event-stream") => {
+                let event_stream =
+                    sse_stream::SseStream::from_byte_stream(response.bytes_stream()).boxed();
+                Ok(
+                    rmcp::transport::streamable_http_client::StreamableHttpPostResponse::Sse(
+                        event_stream,
+                        session_id,
+                    ),
+                )
+            }
+            // A 2xx body that is neither JSON nor SSE (or carries no content
+            // type) is unexpected for a JSON-RPC request — the response-less
+            // 202/204 acks are already handled above. Surface it as an error
+            // like rmcp's reference client rather than silently reporting
+            // `Accepted`, which would drop the real response and hang the
+            // request until it times out.
+            other => Err(
+                rmcp::transport::streamable_http_client::StreamableHttpError::UnexpectedContentType(
+                    other.map(|s| s.to_string()),
+                ),
+            ),
+        }
+    }
+}
 
 /// MCP client for managing connections to MCP servers
 pub struct McpManager {
@@ -298,7 +593,7 @@ impl McpManager {
     ) -> Result<usize, BuilderError> {
         debug!("  Creating SSE client for: {}", url);
 
-        let transport = crate::mcp::sse::SseTransport::connect(url, headers)
+        let transport = crate::mcp_sse::SseTransport::connect(url, headers)
             .await
             .map_err(BuilderError::SseTransport)?;
 
@@ -990,6 +1285,357 @@ impl Default for McpManager {
     }
 }
 
+/// Macro to create unique MCP tool structs for each tool name
+/// This is required because Rig's Tool trait requires each tool to have a unique static NAME
+macro_rules! create_mcp_tool_struct {
+    ($struct_name:ident, $tool_name:literal) => {
+        #[derive(Clone)]
+        pub struct $struct_name {
+            pub tool_name: String,
+            pub server_name: String,
+            pub client: Arc<McpClient>,
+            pub tool_definition: ToolDefinition,
+        }
+
+        impl $struct_name {
+            pub fn new(
+                tool_name: String,
+                server_name: String,
+                client: Arc<McpClient>,
+                mcp_tool: rmcp::model::Tool,
+                sanitize_schemas: bool,
+            ) -> Option<Self> {
+                // Use centralized conversion method
+                let tool_definition =
+                    McpManager::convert_tool_to_rig_definition(&mcp_tool, sanitize_schemas)?;
+                Some(Self {
+                    tool_name,
+                    server_name,
+                    client,
+                    tool_definition,
+                })
+            }
+        }
+
+        impl RigTool for $struct_name {
+            const NAME: &'static str = $tool_name;
+
+            type Error = ToolExecutionError;
+            type Args = Value;
+            type Output = String;
+
+            async fn definition(&self, _prompt: String) -> ToolDefinition {
+                self.tool_definition.clone()
+            }
+
+            async fn call(&self, args: Self::Args) -> Result<Self::Output, Self::Error> {
+                debug!(
+                    "{}::call - Tool: {}, Server: {}",
+                    stringify!($struct_name),
+                    self.tool_name,
+                    self.server_name
+                );
+                info!(
+                    "Calling tool '{}' on server '{}'",
+                    self.tool_name, self.server_name
+                );
+
+                let arguments = match args {
+                    Value::Object(map) => map.into_iter().collect::<HashMap<String, Value>>(),
+                    _ => {
+                        return Err(ToolExecutionError::from(
+                            "Invalid arguments: expected JSON object",
+                        ));
+                    }
+                };
+
+                debug!("  Arguments: {:?}", arguments);
+
+                match self
+                    .client
+                    .call_tool(&self.tool_name, arguments, None)
+                    .await
+                {
+                    Ok(response) => {
+                        debug!("  Response: {}", response);
+                        let response_summary =
+                            crate::mcp_tool_execution::preview_response(&response, 200);
+                        info!("Tool '{}' completed: {}", self.tool_name, response_summary);
+                        Ok(response)
+                    }
+                    Err(e) => {
+                        let error_msg = format!("MCP tool call failed: {}", e);
+                        info!("❌ Tool '{}' failed: {}", self.tool_name, e);
+                        Err(ToolExecutionError::from(error_msg))
+                    }
+                }
+            }
+        }
+    };
+}
+
+// Create unique tool structs for each known Mezmo tool
+create_mcp_tool_struct!(ExportLogsRelativeTimeTool, "export_logs_relative_time");
+create_mcp_tool_struct!(ExportLogsTimeRangeTool, "export_logs_time_range");
+create_mcp_tool_struct!(
+    AnalyzeLogsRelativeTimeTool,
+    "analyze_logs_for_root_cause_relative_time"
+);
+create_mcp_tool_struct!(
+    AnalyzeLogsTimeRangeTool,
+    "analyze_logs_for_root_cause_time_range"
+);
+create_mcp_tool_struct!(GetCurrentTimeTool, "get_current_time");
+create_mcp_tool_struct!(GetPipelineTool, "get_pipeline");
+create_mcp_tool_struct!(ListPipelinesTool, "list_pipelines");
+
+/// Generic fallback for unknown tools - A Rig-compatible tool wrapper for MCP client tools
+#[derive(Clone)]
+pub struct StreamableHttpMcpTool {
+    pub tool_name: String,
+    pub server_name: String,
+    pub client: Arc<McpClient>,
+    pub tool_definition: ToolDefinition,
+}
+
+impl StreamableHttpMcpTool {
+    pub fn new(
+        tool_name: String,
+        server_name: String,
+        client: Arc<McpClient>,
+        mcp_tool: rmcp::model::Tool,
+        sanitize_schemas: bool,
+    ) -> Option<Self> {
+        // Use centralized conversion method from McpManager
+        let tool_definition =
+            McpManager::convert_tool_to_rig_definition(&mcp_tool, sanitize_schemas)?;
+
+        Some(Self {
+            tool_name,
+            server_name,
+            client,
+            tool_definition,
+        })
+    }
+}
+
+impl RigTool for StreamableHttpMcpTool {
+    const NAME: &'static str = "streamable_http_mcp_tool";
+
+    type Error = ToolExecutionError;
+    type Args = Value;
+    type Output = String;
+
+    async fn definition(&self, _prompt: String) -> ToolDefinition {
+        self.tool_definition.clone()
+    }
+
+    async fn call(&self, args: Self::Args) -> Result<Self::Output, Self::Error> {
+        debug!(
+            "StreamableHttpMcpTool::call - Tool: {}, Server: {}",
+            self.tool_name, self.server_name
+        );
+        info!(
+            "Calling HTTP streamable tool '{}' on server '{}'",
+            self.tool_name, self.server_name
+        );
+
+        // Convert Value to HashMap<String, Value> for the streamable client
+        let arguments = match args {
+            Value::Object(map) => map.into_iter().collect::<HashMap<String, Value>>(),
+            _ => {
+                return Err(ToolExecutionError::from(
+                    "Invalid arguments: expected JSON object",
+                ));
+            }
+        };
+
+        debug!("  Arguments: {:?}", arguments);
+
+        // Call the streamable HTTP client
+        match self
+            .client
+            .call_tool(&self.tool_name, arguments, None)
+            .await
+        {
+            Ok(result) => {
+                debug!("  Tool execution successful");
+                let response_summary = crate::mcp_tool_execution::preview_response(&result, 200);
+                info!(
+                    "HTTP streamable tool '{}' completed: {}",
+                    self.tool_name, response_summary
+                );
+                Ok(result)
+            }
+            Err(err) => {
+                info!(
+                    "❌ HTTP streamable tool '{}' failed: {}",
+                    self.tool_name, err
+                );
+                Err(ToolExecutionError::from(format!(
+                    "Tool execution failed: {err}"
+                )))
+            }
+        }
+    }
+}
+
+impl McpManager {
+    /// Get Rig-compatible tools for streamable HTTP MCP clients
+    pub fn get_streamable_http_tools(&self) -> Vec<StreamableHttpMcpTool> {
+        let mut tools = Vec::new();
+
+        for (server_name, client) in &self.streamable_clients {
+            if let Some(server_tools) = self.streamable_tools.get(server_name) {
+                debug!(
+                    "Processing {} streamable HTTP tools for server: {}",
+                    server_tools.len(),
+                    server_name
+                );
+
+                for mcp_tool in server_tools {
+                    if let Some(rig_tool) = StreamableHttpMcpTool::new(
+                        mcp_tool.name.to_string(),
+                        server_name.clone(),
+                        Arc::new(client.clone()),
+                        mcp_tool.clone(),
+                        self.sanitize_schemas,
+                    ) {
+                        tools.push(rig_tool);
+                    }
+                    // If None, tool was rejected due to invalid schema - already logged as warning
+                }
+            } else {
+                debug!("No tools found for streamable server: {}", server_name);
+            }
+        }
+
+        debug!(
+            "Created {} Rig-compatible tools from streamable HTTP clients",
+            tools.len()
+        );
+        tools
+    }
+}
+
+/// Create a fallback tool with a unique name to avoid Claude API "Tool names must be unique" errors
+///
+/// Returns None if the tool has an invalid schema and should be rejected.
+pub fn create_fallback_http_tool(
+    unique_tool_name: String,
+    original_tool_name: String,
+    server_name: String,
+    client: Arc<McpClient>,
+    mcp_tool: rmcp::model::Tool,
+    sanitize_schemas: bool,
+) -> Option<FallbackHttpMcpTool> {
+    FallbackHttpMcpTool::new(
+        unique_tool_name,
+        original_tool_name,
+        server_name,
+        client,
+        mcp_tool,
+        sanitize_schemas,
+    )
+}
+
+/// Fallback tool with dynamic unique naming to avoid Claude API conflicts
+#[derive(Clone)]
+pub struct FallbackHttpMcpTool {
+    pub unique_name: String,
+    pub original_tool_name: String,
+    pub server_name: String,
+    pub client: Arc<McpClient>,
+    pub tool_definition: ToolDefinition,
+}
+
+impl FallbackHttpMcpTool {
+    pub fn new(
+        unique_name: String,
+        original_tool_name: String,
+        server_name: String,
+        client: Arc<McpClient>,
+        mcp_tool: rmcp::model::Tool,
+        sanitize_schemas: bool,
+    ) -> Option<Self> {
+        // Use centralized conversion method from McpManager
+        let mut tool_definition =
+            McpManager::convert_tool_to_rig_definition(&mcp_tool, sanitize_schemas)?;
+
+        // Override the name with our unique name to avoid conflicts
+        tool_definition.name = unique_name.clone();
+
+        Some(Self {
+            unique_name,
+            original_tool_name,
+            server_name,
+            client,
+            tool_definition,
+        })
+    }
+}
+
+impl RigTool for FallbackHttpMcpTool {
+    const NAME: &'static str = "fallback_http_mcp_tool";
+
+    type Error = ToolExecutionError;
+    type Args = Value;
+    type Output = String;
+
+    async fn definition(&self, _prompt: String) -> ToolDefinition {
+        self.tool_definition.clone()
+    }
+
+    async fn call(&self, args: Self::Args) -> Result<Self::Output, Self::Error> {
+        debug!(
+            "FallbackHttpMcpTool::call - Original Tool: {}, Unique Name: {}, Server: {}",
+            self.original_tool_name, self.unique_name, self.server_name
+        );
+        info!(
+            "Calling fallback HTTP tool '{}' (original: '{}') on server '{}'",
+            self.unique_name, self.original_tool_name, self.server_name
+        );
+
+        // Convert Value to HashMap<String, Value> for the streamable client
+        let arguments = match args {
+            Value::Object(map) => map.into_iter().collect::<HashMap<String, Value>>(),
+            _ => {
+                return Err(ToolExecutionError::from(
+                    "Invalid arguments: expected JSON object",
+                ));
+            }
+        };
+
+        debug!("  Arguments: {:?}", arguments);
+
+        // Call the streamable HTTP client using the original tool name
+        match self
+            .client
+            .call_tool(&self.original_tool_name, arguments, None)
+            .await
+        {
+            Ok(result) => {
+                debug!("  Fallback tool execution successful");
+                let response_summary = crate::mcp_tool_execution::preview_response(&result, 200);
+                info!(
+                    "Fallback HTTP tool '{}' completed: {}",
+                    self.unique_name, response_summary
+                );
+                Ok(result)
+            }
+            Err(err) => {
+                info!(
+                    "❌ Fallback HTTP tool '{}' failed: {}",
+                    self.unique_name, err
+                );
+                Err(ToolExecutionError::from(format!(
+                    "Fallback tool execution failed: {err}"
+                )))
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1051,6 +1697,78 @@ mod tests {
         assert!(
             snapshot[0].reason.is_some(),
             "failed server should carry a reason"
+        );
+    }
+
+    /// `post_message` must (a) forward the negotiated `mcp-session-id` header on
+    /// every post after `initialize`, and (b) treat a `202 Accepted` /
+    /// `204 No Content` response as `Accepted` *without* parsing the (empty)
+    /// body — even when the server tags that empty body `application/json`
+    /// (FastMCP does).
+    ///
+    /// Both are regression guards for the `notifications/initialized` post:
+    /// dropping the session id makes the server reject it (FastMCP 400, rmcp
+    /// 422); parsing the empty 202 body as JSON-RPC fails to deserialize. Either
+    /// bug collapses into a generic "channel closed" and the whole server is
+    /// recorded as `Failed`, so no tools are available.
+    #[tokio::test]
+    async fn post_message_forwards_session_id_and_accepts_empty_202() {
+        use std::sync::{Arc, Mutex};
+        use tokio::io::{AsyncReadExt, AsyncWriteExt};
+        use tokio::net::TcpListener;
+
+        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
+        let addr = listener.local_addr().unwrap();
+
+        // Capture the raw request bytes the server received so the test can
+        // assert the session-id header was actually sent on the wire.
+        let seen_request = Arc::new(Mutex::new(String::new()));
+        let seen_for_server = Arc::clone(&seen_request);
+
+        let server = tokio::spawn(async move {
+            let (mut sock, _) = listener.accept().await.unwrap();
+            let mut buf = vec![0u8; 4096];
+            let n = sock.read(&mut buf).await.unwrap();
+            *seen_for_server.lock().unwrap() = String::from_utf8_lossy(&buf[..n]).to_string();
+            // FastMCP-style ack: 202 Accepted, application/json content-type,
+            // and an empty body.
+            sock.write_all(
+                b"HTTP/1.1 202 Accepted\r\ncontent-type: application/json\r\ncontent-length: 0\r\n\r\n",
+            )
+            .await
+            .unwrap();
+            let _ = sock.flush().await;
+        });
+
+        let client = CustomHttpClient::from_reqwest(reqwest::Client::new());
+        let uri: Arc<str> = format!("http://{addr}/mcp").into();
+        let message: rmcp::model::ClientJsonRpcMessage = serde_json::from_value(
+            json!({"jsonrpc": "2.0", "method": "notifications/initialized"}),
+        )
+        .unwrap();
+        let session: Arc<str> = "test-session-abc123".into();
+
+        let result = client
+            .post_message(uri, message, Some(Arc::clone(&session)), None)
+            .await;
+
+        server.await.unwrap();
+
+        // (b) Empty 202 → Accepted, not a deserialize error.
+        assert!(
+            matches!(
+                result,
+                Ok(rmcp::transport::streamable_http_client::StreamableHttpPostResponse::Accepted)
+            ),
+            "empty 202 should yield Accepted, got {result:?}"
+        );
+
+        // (a) The session id was forwarded as the mcp-session-id header.
+        let request = seen_request.lock().unwrap();
+        let lowered = request.to_lowercase();
+        assert!(
+            lowered.contains("mcp-session-id: test-session-abc123"),
+            "post_message must forward the session id header; request was:\n{request}"
         );
     }
