+++ crates/aura/src/mcp_response.rs
